+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+- Add support from creating an endpoint using the new [swift-http-types] package
+
+[swift-http-types]: https://github.com/apple/swift-http-types
+
 ## [v2.0.1] - 2023-06-19
 
 - Improve how documentation is rendered on the Swift Package Index
@@ -63,6 +69,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 Initial release of the library
 
+[Unreleased]: https://github.com/rhysforyou/Superhighway/compare/2.0.1...HEAD
 [v2.0.1]: https://github.com/rhysforyou/Superhighway/compare/2.0.0...2.0.1
 [v2.0.0]: https://github.com/rhysforyou/Superhighway/compare/1.1.0...2.0.0
 [v1.1.0]: https://github.com/rhysforyou/Superhighway/compare/1.0.0...1.1.0